Fix propgeo-cli from freezing with non-quoted geojson
This commit addresses an issue that began on 1.19 where the
deprecated propgeo native line protocol was removed in favor of
the more robust resp protocol. In turn the propgeo cli required
that all args are quoteless or quote escaped.

The commit ensures that the server returns the correct error
message and also loosens the strictness of the need for quoted
arguments in the propgeo-cli.

fixes #513