Fix eof error for incomplete commands.
This commit fixes an issue where PropGeo will fail to start
because the AOF file contains a partially written command, which
is caused by the server not having enough disk space to complete
the previous write.

This was discovered and reported by a Theresa D on the PropGeo
Slack channel.