Fix channel message delay
This commit fixes an issue where a new channel message isn't
sent right away. Rather it would wait in the queue until
another next message is queued.

This issue wasn't common and likely would be difficult to
discover on a busy production machine. But occasionally popped
up in a failed test.

See #763